Last little features (#89)
* Add grid lines for moving-- initial commit

* Add headers for TBW sections; refactor TBW

* Implemented Functionality of Trigger actions, Switched up Trigger Propery window layout, Added Collectable Component

* Added jump trigger, made CCollectable display a drop of game objects with the CText component, Fixed respawn of Entity after it's Chealth has been made less than 0

* Fix clone bug

* Add collectible button-- template, initial commit

* Fix responsive buttons & refactor

* Refactor Grid stuff

* More TBW refactoring

* Add CHealth to player button

* Cmake

* Need work on scenes

* cmake

* Removal of Corrupted Last_scene.json, Fixing of Editor States after selecting entity from explorer window, Fixing Orign reset

* Added Reset functions to CHealth, CCollectable, and in progress of fixing deserialization of CTouchTrigger

* Fixed Trigger loading, and reseting of health and collectables

* Finished merge, update gator physics to remove old entities when changing scene

* Added Reset to CText score, fixed animation bug that was a result of the merge, rounded score of CText

* Moved NFD fetch behind SMFL to avoid Cmake build issue with GitHub

* CMake

* CMake

* Delete SceneMenu

* Delete Scene

* Update Insert Player button

* Delete Scene_Play and Scene_Old (moved into GatorEngine)

---------

+++ src/editor/Editor.h
@@ -46,9 +46,7 @@ public:
 	// Grid properties
 	static bool show_grid_;
 	static bool snap_to_grid_;
-	static float grid_size_;
-
-
+	static int grid_size_;
 private:
 	// List of all windows in the Editor
 	std::vector<std::unique_ptr<UIWindow>> windows_;
